requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.5
django-3-jet==1.0.8
django-comments-xtd==2.9.1
django-contrib-comments==2.1.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-rest-passwordreset==1.1.0
django-rest-resetpassword==0.1.2
django-rest-swagger==2.2.0
django-utils-six==2.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
docutils==0.17.1
drf-yasg==1.20.0
gunicorn==20.0.4
idna==2.10
importlib-metadata==3.4.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
Markdown==3.3.3
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==20.9
Pillow==8.1.2
psycopg2-binary==2.8.6
PyJWT==2.0.1
pyparsing==2.4.7
pytz==2020.5
requests==2.25.1
ruamel.yaml==0.17.2
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
Unidecode==1.2.0
uritemplate==3.0.1
urllib3==1.26.4
zipp==3.4.0
